% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ble}
\alias{ble}
\title{Poids au mille grains de ble}
\format{
\subsection{\code{ble}}{

A data frame with 12 rows and 2 columns :
\describe{
\item{variete}{Factor, Name of variety}
\item{PMG}{Poids au mille grains, thousand grains weight}
}
}
}
\source{
Data generated by Julien Bousquet (2025)
}
\usage{
ble
}
\description{
Illustration of Completely Random Design
3 different varieties of wheat are planted
and yield are measured in  thousand grain weight.
}
\examples{
data(ble)
}
\keyword{datasets}
