% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{cochenille}
\alias{cochenille}
\title{Infestation de cycas par des cochenilles}
\format{
\subsection{\code{cochenille}}{

A data frame with 30 rows and 3 columns :
\describe{
\item{Diff}{Difference of number of mealybugs before-after treatment}
\item{trt}{Faactor, treatment name}
\item{plant}{Factor, identification of the plant (numbered cycas) on which experiment have been done.}
}
}
}
\source{
Data shared by Scott Smith in A First Course in
Design and Analysis of Experiments of Gary W. Oehlert
}
\usage{
cochenille
}
\description{
Illustration of Randomized Complete Bloc Design.
}
\examples{
data(cochenille)
X <- aggregate(Diff~trt+plante, mean, data=cochenille)
}
\keyword{datasets}
