% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{DC}
\alias{DC}
\title{Drug Concentration for different way of administration}
\format{
\subsection{\code{DC}}{

A data frame with 9 rows and 4 columns :
\describe{
\item{conc}{Area under time-concentration curve}
\item{sujet}{Subject name}
\item{periode}{Periode number}
\item{adm}{Administration mode}
}
}
}
\source{
Data proposed by  Gary W. Oehlert in A First Course in
Design and Analysis of Experiments
}
\usage{
DC
}
\description{
Illustration of Latin Square Design.
}
\examples{
data(DC)
}
\keyword{datasets}
