% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{detergent}
\alias{detergent}
\title{Detergent efficiency}
\format{
\subsection{\code{detergent}}{

A data frame with 36 rows and 3 columns :
\describe{
\item{trt}{Name of treatment}
\item{nb.plates}{Number of plate cleaned before foam disparition}
\item{session}{Number of session}
}
}
}
\source{
Data proposed by  Gary W. Oehlert in A First Course in
Design and Analysis of Experiments
}
\usage{
detergent
}
\description{
Illustration of Balanced Incomplete Block Design
Three workers cleaned plate at the same rythm with different detergent
solutions. They count the number of plates they cleaned before
foam disparition. Every worker goes at the same pace
in a session. Speed changed from a session to another.
}
\examples{
data(detergent)
}
\keyword{datasets}
