% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{DrugConcentration}
\alias{DrugConcentration}
\title{Drug Concentration for different way of administration}
\format{
\subsection{\code{DrugConcentration}}{

A data frame with 9 rows and 4 columns :
\describe{
\item{conc}{Area under time-concentration curve}
\item{sujet}{Factor, subject name}
\item{periode}{Factor, periode number}
\item{adm}{Factor, administration mode}
}
}
}
\source{
Data proposed by  Gary W. Oehlert in A First Course in
Design and Analysis of Experiments
}
\usage{
DrugConcentration
}
\description{
Illustration of Latin Square Design.
}
\examples{
data(DrugConcentration)
}
\keyword{datasets}
