% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{orge}
\alias{orge}
\title{Germination de l'orge}
\format{
\subsection{\code{orge}}{

A data frame with 30 rows and 3 columns :
\describe{
\item{germination}{Integer, Nombre de graines germees}
\item{eau}{Factor, quantite d'eau en ml}
\item{age}{Factor, age de la graine en nombre de semaines}
}
}
}
\source{
Data proposed by  Gary W. Oehlert in A First Course in
Design and Analysis of Experiments
}
\usage{
orge
}
\description{
Illustration of Factorial Design
Etude de germination de 100 graines d'orge
reparties en 30 lots.
}
\examples{
data(orge)
}
\keyword{datasets}
